third_party/WebKit/LayoutTests/editing/style/apple-style-editable-mix.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
test(() => assert_selection(
    [
        '<div contenteditable>',
            '^abc',
            '<div contenteditable="false">',
                'def',
                '<div contenteditable>xyz</div>',
                'ghi',
            '</div>',
            'def|',
        '</div>',
    ].join(''),
    'bold',
    [
        '<div contenteditable>',
            '<b>^abc</b>',
            '<div contenteditable="false">',
                'def',
                '<div contenteditable><b>xyz</b></div>',
                'ghi',
            '</div>',
            '<b>def|</b>',
        '</div>',
    ].join('')),
    'Apply style to mixed editable contents');
</script>
